requirements_dev.txt
-r requirements.txt
attrs==19.3.0
pytest==5.3.3
pytest-django==3.8.0
pytest-celery==0.0.0a1
splinter==0.13.0
pytest-splinter==2.0.1
responses==0.10.9
flake8==3.7.9
pydocstyle==5.0.2
pylint==2.4.4
pylint-django==2.0.13
git+https://github.com/IATI/wagtail-factories.git@35e1bbe135acec42c47f9e0ddb73faf20f2cce32#egg=wagtail-factories
factory-boy==2.11.0
sqlparse==0.3.0
